+++ tests/test_parameters.py
@@ -522,7 +522,7 @@ def test_integrated_show_params_option(invoke, create_config):
             "✘",
             "✓",
             "SHOW_PARAMS_CLI_VERBOSITY",
-            "<LogLevel.WARNING: 30>",
+            "'WARNING'",
             "'DeBuG'",
             "COMMANDLINE",
         ),